Update patch set 1
Patch Set 1: Code-Review+1

this makes sense to me.

i don't have the history but the 1.0.14 release is suppose to roughly be the changes in 1.1.0 and 1.2.0 (minus the requirements changes)

visually the 1.1.0 requirements are less than existing kilo requirements. i assume if we cap kilo at 1.1.0 we will now have to backport all changes since. which is fine but if there's a way to minimise what we need to backport the better)

just for reference, we're currently targetting 1.5.0+ as our liberty candidate.

Patch-set: 1