Align pyngus with oslo.messaging stable
Oslo messaging stable is using pyngus>=1.3,<2 because previous versions
don't work in the gate. This was not picked by the tests because it, for
now, installs that using a different requirements file rather than the
common one.

This commit aligns this requirements with what's in oslo.messaging
stable/kilo in the amqp1-requirements.txt file.

Change-Id: I8b7750c6a5ad06d9a233c58d1e8c2d18e5f63ae5