edit-constraints deal better with libraries with underscores
Problem showed up with django_openstack_auth job that installs
from source. The change to devstack to verify LIBS_FROM_GIT
kicked in (Iffef2007f99a0e932b68c4c897ebbfb748cac2b4) and
starting failing the job. The problem was that the
edit_constraint was not fixing the django-openstack-auth
line in upperconstraints.txt as the setup.cfg had
name = django_openstack_auth. Note that when things finally
get installed setup_tools' safe_name kicks in to
standardize names and for example glance_store will show
up as glance-store in pip freeze. So we should be able
to treat this situation better and allow constraints to
have the safe names (using dashes).

Change-Id: Ibaa22657a2cf2c0ad96dbd0b9bc43cdafe6a1d56