Pin pbr<1.8 which ignores Sphinx warnerrors
By setting pbr.warnerros = True, Sphinx will fail the build on any
warnings ensuring the documentation is properly maintained.

Due to a bug in pbr 1.8.0, the setting is not recognized when using
Sphinx < 1.3.1 https://bugs.launchpad.net/pbr/+bug/1496882

Since the global-requirements.txt has Sphinx<1.3, it is hit by the bug
and that cause gates to leak Sphinx warnings in repos.

Another way would be to bump Sphinx>=1.3.1 and maybe pbr>=1.8, but that
seems to be a terrible idea.

Restrict pbr to not be above 1.8.
When the bug is fixed, pbr 1.8.0 will need to be blacklisted until
Sphinx is bumped to >=1.3.1.

Change-Id: I3e0a9b625d99d2beb117554948ea738a8fb9047b