Update patch set 1
Patch Set 1: Code-Review-2

We're in a requirements freeze right now. Unless another OpenStack project needs the features present in this release of nova client, we should not update the lower requirement. The constraints file was updated in another patch syncing the constraints for all of our releases, so that part of this change isn't needed.

Patch-set: 1
Label: Code-Review=-2