Update patch set 2
Patch Set 2:

(1 comment)

This is the diff between 1.1.9 and 1.2.0:

https://github.com/openstack/django_openstack_auth/compare/1.1.9...1.2.0

There are a few things of note:

1. py26 support is dropped: 6e5dae01adf2ab3eb7bb9b9afa08f8ba95d8eafc

2. keystone auth plugins are used: 01e0abc17d5cf451e0e7556a90ad6fb2de2c78a5

3. moved to hacking 0.10: 31b7e7c103fb9c5dd16f2b0f67666485c43355d9

--

In kilo we dropped py26 support in the server projects but not necessarily the libraries.  Given the change happened during kilo though it seems like this would be OK.

The hacking change is OK given g-r on stable/kilo.

I'm not sure how compatible the keystone auth stuff is.  That change was also made in kilo though so it seems like it'd be OK.

--

I'd like David Lyle to ack this before we approve.

Patch-set: 2