Update patch set 2
Abandoned

This patch was merely for demo purposes.  I used it as a dependency of openstack/mistral patch https://review.openstack.org/#/c/224761/  to make it pass.

Per discussion with lifeless, we need a better patch in pbr that honors pbr.warnerrors for both Sphinx 1.2.x and Sphinx 1.3.x.  Then a new pbr release and a blacklist of pbr 1.8.0.

Patch-set: 2
Status: abandoned