Update patch set 2
Patch Set 2:

Well, new release of pbr.

We could raise the buildtime minimum for openstack as a whole, but since we already use the highest version, and builds won't *fail* on lower versions - in fact, they will build correctly if the docs are ok, its only CI thats impacted - there's little reason to do so.

The runtime API isn't the one in question here - the reason the tests were able to pass is because the buildtime API is also changed when a dep is installed.

Patch-set: 2