Bump pbr to 1.8
Currently, we have global-requirements.txt with pbr >= 1.6, but
setup.py has >= 1.8. As a consequence, many packages have a
simply wrong requirements.txt, as the proposal bot gives bad
advices. For example:
https://review.openstack.org/#/c/224618/

The consequence can be that pbr gets downloaded during the
clean target, which is bad (this fails package building in
some cases, like with a pbuild / sbuild).

So, the best way forward to fix this bad situation, is IMO
to bump pbr to 1.8 right now. I don't think it's a problem for
downstream distributions anyway (it's not for me in Debian at
least...).

Change-Id: I7a70beaf5312f2436ca182fcaaadabff7cf9cf48