Update patch set 1
Patch Set 1: Code-Review-2

This is coming really late. Now that we have stable branches for a lot of projects, updating the requirement will mean backporting requirements changes to all of the projects that use the library directly. A simple grep shows 36 projects under openstack/* using tempest-lib, and although a lot of those haven't branched yet many have.

Is the work that would be blocked by this for liberty, or mitaka?

Patch-set: 1