Update patch set 1
Patch Set 1:

Updating this requirement will also trigger updates for 5 other projects:

  $ grep designate */*requirements*.txt
  designate-dashboard/requirements.txt:python-designateclient>=1.0.0
  designate/requirements.txt:python-designateclient>=1.0.0
  heat/requirements.txt:python-designateclient>=1.0.0
  rally/requirements.txt:python-designateclient>=1.0.0
  requirements/global-requirements.txt:python-designateclient>=1.0.0
  searchlight/requirements.txt:python-designateclient>=1.0.0
  trove/requirements.txt:python-designateclient>=1.0.0

Rather than force all of them to spin new releases, we are going to wait until we have the requirements repository branched for liberty. Then we can land this patch in master, and consider a backport of the requirement change as a separate case.

We need to be more careful about raising the minimum required versions of our dependencies as we adopt new features.

Patch-set: 1