Add freezegun to global requirements
Source: https://github.com/spulec/freezegun

PyPI: https://pypi.python.org/pypi/freezegun

This allows us to reproduce race conditions by controlling the clock in
third party libraries (such as PyPI/cryptography) which do not depend on
oslo_utils.timeutils (and thus cannot be controlled by timeutils' clock
freezing methods).

It allows you to universally freeze the clock when accessed via:

- datetime.datetime.now()
- datetime.datetime.utcnow()
- datetime.date.today()
- time.time()
- time.localtime()
- time.gmtime()
- time.strftime()

Using, for example, a context manager:

  with freezegun.freeze_time(datetime.datetime.utcnow()):
    pass

Consuming projects
==================

This is a proposed requirement of keystone:

  https://review.openstack.org/#/c/227995/

I also believe it could be used to supplant oslo time_utils
limited implementation of the same behavior.

Is the library actively maintained?
===================================

Yes, the library is actively maintained by a small group of
contributors:

  https://github.com/spulec/freezegun/commits/master

The latest release was August 2015.

Is the library good code?
=========================

The actual library is small, and although mostly undocumented, is well
structured and easy to follow:

  https://github.com/spulec/freezegun/blob/master/freezegun/api.py

The test suite is quite exhaustive and includes plenty of negative
tests, for example:

  https://github.com/spulec/freezegun/blob/master/tests/test_datetimes.py

Is the library python 3 compatible?
===================================

freezegun claims compatibility with at least Python 3.3. It seems to
work in Python 3.4 for the use cases I've tested.

Is the library license compatible?
==================================

freezegun's LICENSE file:

  https://github.com/spulec/freezegun/blob/master/LICENSE

Is the library already packaged in the distros we target?
=========================================================

freezegun is packaged for Ubuntu:

  http://packages.ubuntu.com/vivid/python-freezegun
  http://packages.ubuntu.com/vivid/python3-freezegun

... and for Fedora:

  https://apps.fedoraproject.org/packages/python-freezegun
  https://apps.fedoraproject.org/packages/python3-freezegun

Is the function of this library already covered by other libraries?
===================================================================

Other than oslo time_util's limited implementation, frozendate is the
other, seemingly less popular and less comprehensive option that I'm
aware of and that's not in openstack/requirements either.

Change-Id: I6daed50a9128f0c6208ed1d432177d2da0613e14
Related-Bug: 1473567