Add sphinx_rtd_theme to enhance doc gate jobs
When it gets to submitting & reviewing patches for the project documentation,
it is very convenient to surf through the docs generated by Sphinx in a special
Gate job. The problem is that Sphinx uses in gates a theme which is different
from that used on ReadTheDocs. So the docs generated in gates look differently
from the docs that we have on ReadTheDocs, while the desired behaviour is to
have the same Sphinx theme both on ReadTheDocs and in gates.

Change-Id: Idd899148184c0490099b2d1b141342d86bcef458