Cap requests at <2.8.0 since it packages urllib3 1.12
We have a cap set on urllib3 at 1.11 and requests 2.8.0, released on
10/6, packages urllib3 1.12, which introduces a conflict.

So cap requests at 2.8.0 to avoid issues with newer urllib3.

Since python-glanceclient depends on requests this will need to sync to
glanceclient on stable/kilo and then we can do a patch version release
of that library.

Partial-Bug: #1476770

Change-Id: I8ff76853220c10565b06ba74a7c75c5a955e871b