Update patch set 3
Patch Set 3:

Reading the coverage patch only os is affected by the fix sys, json, socket, re, etc are all still possibly in this boat. It seems odd to consider this fixed in that case when any mocks of those other stdlib modules would have to be defensive as well.

Do we really want to exclude every version of coverage that trips up any of these potentially mocked items? I really do think that our tests have to be smarter if we expect coverage to cover the tests.

Patch-set: 3