Update patch set 5
Patch Set 5:

I've started a thread on this on -dev.

The practical issue here is that this won't fix the problem in a systematic way, and will give us pip version conflicts and force releases of *everything* we produce that consumes either requests or urllib3 anytime we upgrade either of those. Thats a really big problem.

Patch-set: 5