Update patch set 3
Patch Set 3:

I agree with you that coverage should add defensive code to all of the imports that are being used in the code it is injecting into the code under test.  That way it would not be impacting the ability to fully test user code.  I communicated this idea to the developer that was working on this fix and I hope that it will be implemented widely in a future update.

I however don't understand how we can effectively make our tests "smarter".  In this case coverage added an import to code they inject into our test code without protecting it from interacting with our code and test code.  The tests this regression caused to fail are valid tests that check to make sure the code properly handles the presence of files, such as security certificates being generated in one case, correctly. 

Until coverage is fully updated to not impact the code under test I see no other option than to exclude versions that have regressions that break existing tests.

Patch-set: 3