Update patch set 1
Patch Set 1:

Tempest's requirements on master will never be able to match with the requirements on stable branches. Tempest only has a master branch, there are no stable branches on tempest. So we only keep tempest master in sync with master g-r.

Even if you were to backport the tempest-lib bump there is no guarantee that any of the other requirements match from master tempest are in sync with what we have in global requirements on a stable branch. Keeping tempest in sync with with global requirements on master is mutually exclusive from being co-installable with a project's stable branches. It's like asking to install stable/juno neutron on the same node as master nova. 

This is why for gate runs (and any stable devstack deployment) we isolate the tempest install in a virtualenv to make sure the requirements don't interfere with each other. If you need to co-install tempest on the same box as a stable branch deployment (which I never recommend since there is no reason to do this as tempest is essentially an api client) you have to use a venv to install tempest.

Patch-set: 1