Update patch set 1
Patch Set 1: Code-Review-1

Tony Breeds, this does not resolve the issue we are seeing in all the trove tests. The issue we have is with oslo.utils version 2.6.0. If there was a new version cut for oslo.utils this would not be a high priority issue for us.

With oslo.utils==2.6.0 trove is seeing the tox py27 tests hang currently. There is a patch i applied to my venv on oslo.utils that resolves this but its not released yet. 

Since oslo.utils was just cut as well with other oslo projects could we exclude the oslo.utils==2.6.0 version as well in the patch[1] if making another patch is not a good idea?

[1] https://review.openstack.org/#/c/235588/

Thanks - Craig

Patch-set: 1