Update patch set 3
Patch Set 3: Code-Review+2

The code change looks good. Since this is going to be a multi-part fix it would be useful to have a bug or something to use for tracking  the work. I'm not sure which project to recommend filing the bug against, since I don't think we have a LP project for the requirements repo.

Patch-set: 3
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2