Add requestsexceptions to global requirements
Source:
https://git.openstack.org/cgit/openstack-infra/requestsexceptions

PyPI: https://pypi.python.org/pypi/requestsexceptions

Import exceptions from potentially bundled packages in requests.

The python requests library bundles the urllib3 library, however, some
software distributions modify requests to remove the bundled library.
This makes some operations, such as supressing the "insecure platform
warning" messages that urllib emits difficult. This is a simple library
to find the correct path to exceptions in the requests library regardless
of whether they are bundled.

Consuming projects
==================

This is used by shade.

This will be a proposed requirement of os-client-config.

Is the library actively maintained?
===================================

Yes, the library is actively maintained by the OpenStack Infra team.

The latest release was November 2015.

Is the library good code?
=========================

The actual library is very small, and although mostly undocumented, is
well structured and easy to follow:

http://git.openstack.org/cgit/openstack-infra/requestsexceptions/tree/requestsexceptions/__init__.py

Is the library python 3 compatible?
===================================

Yes.

Is the library license compatible?
==================================

freezegun's LICENSE is Apache 2.0:

Is the library already packaged in the distros we target?
=========================================================

No, but there is no reason packaging it will be hard.

Is the function of this library already covered by other libraries?
===================================================================

Nope.

Change-Id: I52249b6d2fe04c49a9f4ed139d7625c890309ca8