Update patch set 1
Patch Set 1:

OK, we don't generally increase minimums for bug fixes. It's not clear if this case is a critical-no-one-should-ever-use-the-old-version situation or not. If the upper-constraints.txt file is up to date, we should be using the newest version already, and if the APIs work for the older versions we shouldn't need to raise the minimums.

Patch-set: 1