Update patch set 3
Patch Set 3:

Today is stable/juno freeze for 2014.2.4, with planned release on 11/19.

How much do we care about getting this in, syncing it to the projects that use it, and then releasing those?

What are we ultimately trying to land that requires this?  A horizon g-r sync?

Patch-set: 3
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0