Update patch set 3
Patch Set 3:

@Matt  Well there are a few Horizon patches blocked by this so *perhaps* it's a little more that just the g-r sync.  I'm suggesting that add a bunch of stuff into stable/juno and then release.

I admit I'm new and naive but it'd be great if the final tag for juno could (with too much hassle) be installed and this facilitates that.

I admit I'm on the fence I'd love to see stable/juno go out and work but who knows what else we'll uncover/break in the process.

Patch-set: 3