Update patch set 2
Patch Set 2:

gnocchi is not run on the same managed release cycle as most other projects. the current master is designed to be api compatible with liberty openstack.

the intent is to be able to support stable/1.3[1] and the future stable/1.4 against liberty openstack

to be honest, i'm not sure why upper-constraints exists. understandably it saves the gate, but as it's not a real upper cap, it's really just hiding potential real world issues. (not trying to re-ignite the long thread on this topic)

[1] https://review.openstack.org/#/c/244689/

Patch-set: 2