Add python consul client
This client library is used to interact with Consul. Consul
driver will be added to Tooz drivers.

Is the library actively maintained?
  Yes, cablehead seems to be the guy who is maintaining
  it. This library is just a wrapper around python requests
  module (https://github.com/cablehead/python-consul).
Is the library good code?
  Yes.
Is the library python 3 compatible?
  Yes.
Is the library license compatible?
  Yes.
Is the library already packaged in the distros we target?
  Yes.
Is the function of this library already covered by other
libraries in global-requirements.txt?
  Nope

Change-Id: I4f57a2f37f268a66b29b81a9401c1a2edba1fe7a