Block oslo.messaging 2.6.0->2.9.0 in Liberty
2.6.0 changed module names for notification drivers in change:

1893c495f6e9ba53c7d93ad99f679f44f66837ef

This broke Neutron's notification system (and maybe other projects).

Paths were updated in Neutron for Mitaka in change:

225663888e71d44e9732951f515da925b8799b74

And a minimum set on oslo.messaging in mitaka g-r.

To fix Neutron in stable/liberty, oslo.messaging now has an alias for
the older messaging driver paths:

15cd99050c6d2714b90059e0faad9f9e3409eaaa

This is released in oslo.messaging 2.10.0. So we need to block all
oslo.messaging releases that don't have the internal driver paths, which
is all releases between 2.6.0 and 2.9.0.

Co-Authored-By: Matt Riedemann <mriedem@us.ibm.com>

Closes-Bug: #1513630

Change-Id: I5d15366a212c6ed21f33bfe011876fc2bb56997c