Update patch set 1
Patch Set 1: Code-Review-2

We should go with this:

https://review.openstack.org/#/c/247276/

As dims pointed out in the bug report, neutron was specifying an internal path within oslo.messaging, which is fair game to rename/move in a minor version release. So let's just fix this in neutron where it was doing the wrong thing to begin with.

Might require a release note in stable/liberty neutron any liberty neutron prior to the backport that is using oslo.messaging>=2.6.0.

Patch-set: 1
Label: Code-Review=-2