Update patch set 2
Patch Set 2:

Spoke with dims, we're going to create an alias to the changed name in oslo.messaging, release that, and then we'll blacklist oslo.messaging 2.6.0->the release that has the alias in stable/liberty g-r.  That should fix neutron in stable/liberty.

And the alias usage will be deprecated so people know to move off it (like in mitaka).

Patch-set: 2