Update patch set 1
Patch Set 1: Code-Review-2

We don't want to do this, because it prevents any project from using the library. Just because we don't need it doesn't mean we should prevent it from being installed. Cliff now checks at runtime whether to use unicodecsv or not, so it's safe to install and ignore it.

https://review.openstack.org/#/c/226966

Patch-set: 1
Label: Code-Review=-2