Update patch set 1
Abandoned

lifeless wrote on another change: "so, I think you've misunderstood things here. The unicodecsv stuff was *not related to the 3.5 issue*. Doug put that in as a dependent change but the actual bug was only the class docstring lookup stuff. So this is fine."

Oh ok. Well, I don't want to spend time on this issue, so I just give up :-p

Patch-set: 1
Status: abandoned