Add BetaMax to global requirements
BetaMax is super useful when coupled with requests as a way to
record a session and set of HTTP requests and then replay the actions
within a testing scenario (without needing access to the live web
resources). This will be used in support of a fixture in KeystoneAuth
and OCC so that we can mock out auth-requests and verify that a change
will not break compatibility with active/live clouds.

Change-Id: I8787954ab5d0e3963e97c160783301b48ab8299f