Update patch set 5
Patch Set 5:

So we can just ignore Jeremy's concerns? the reason we turned the job to non-voting is because we didn't know the time scales on getting the non-python3 dependencies python3 compatible, and because they aren't in global requirements yet the check-requirements job was blocking our gate. If that is no longer a concern then I'm more than happy to turn it back voting and to get this patch merged.

Patch-set: 5