bump uppper constraints of keystone libraries
in preparation of mitaka the keystone team will be releasing new
versions of all its libraries. py26 support has been dropped on
each and result in a major version bump for all libraries.
bump upper-constraints accordingly.

Due to a bug in version 2.0.0 of pycadf, we are excluding that
release. Newer versions, such as 2.0.1 are fine.

Related-Bug: 1521844

Change-Id: Ibc70ea0ec54d5edbd420cf7c82e6ae8312e887e5
Depends-On: I1df68d34e399a15f416b238d55d456b829208d9f