Update patch set 2
Patch Set 2: Code-Review-1

We are dropping support for python2.6 in master of the clients and libs right nowish. Prefer that we don't add this because we will just have to remove it in a day or two.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1