Update patch set 2
Patch Set 2:

If we do this cap, it means we can't test changes of anything that depends on o.middleware >= 3 against liberty. IIRC all the deps on o.middleware are from servers, so that should be ok.

BUT, it also means we can't do any test where there's a conflicting dep between o.middleware <3 and e.g. newer oslo.config etc - so we're going to have to be super careful not to do *any other* backward incompat breaks in the dependencies of oslo.middleware, otherwise we'll lose the ability to test backwards compat entirely.

Really, it would be better to have kept compat until liberty eol, where this problem wouldn't arise.

Patch-set: 2