Update patch set 3
Patch Set 3: Code-Review+1

Yeah, api-paste.ini is config, IMHO, and we have LONG said that we will not break people's configs from release to release. That's a pretty low bar, and an embarrassing one for us to not be able to consistently hit.

I agree that all projects would fail to meet the supports-upgrade tag if we're breaking config, and by the rules of that tag, nova would also not qualify for the rolling-upgrade tag.

Capping this for the benefit of our users seems like far and away the best choice to me.

Patch-set: 3
Reviewer: Gerrit User 4393 <4393@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1