Update patch set 3
Patch Set 3:

@Dan, to reiterate what was discussed in the release channel this morning, theoretically people deploying openstack now (packaging it, pip installing it, whatever), are supposed to rely on the upper-constraints file for knowing the max safe versions to use.

That's discussed here:

http://lists.openstack.org/pipermail/openstack-dev/2015-October/077162.html

I personally think that's unrealistic as it requires installing python packages from openstack differently from any other python package I'm aware of - you have to pull part of your dependencies from another repo, which is counter-intuitive to just git cloning and pip installing a single project.

But that's where we are right now, so before we just start capping things on stable (which I'm still more or less in favor of), we have to sort that out.  Also, if we cap oslo.middleware here, we also have to cap it's dependencies too else they'll change and break the capped oslo.middleware, which got us into this mess with stable branch caps in the first place.

Patch-set: 3