Update patch set 1
Patch Set 1: Code-Review-1

This is pointless since upper-constraints (and requirements lists in general) are use _by_ tox (via pip), and so cannot be used to determine what version of tox to use. Our job workers always have the latest version of tox preinstalled, so any alteration to the version used has to happen in the image build process. In this case we've simply rolled back the impacted images and halted further image rebuilds while we wait for tox to address the issue with a newer release.

Patch-set: 1
Label: Code-Review=-1