Update patch set 1
Patch Set 1:

The upper-constraints.txt file is autogenerated. Our automation is not designed to support adding entries to it which are not derived from things either directly or transitively referenced in global-requirements.txt.

You would need this to be part of global-requirements.txt first, and we would need to rework all our jobs to bootstrap tox (and do so respecting the constraints file!) before using tox to install all the actual requirements of the job. I think it's undesirable to further complicate our application of constraints, which is still in its infancy, to address this particular corner case (tox, and presumably other python-based command line tools used in our CI such as pip, virtualenv, bindep, zuul-cloner, ...).

Patch-set: 1