Update patch set 1
Patch Set 1: Code-Review-1

We don't cap requirements in stable now, or there should be exceptional reasons for doing so. We've had similar issues with oslo libraries lately (renaming or removing things) and we've made the oslo libraries add compat shims back in which will be removed after liberty-eol.

Is there a bug for this breakage?

Patch-set: 1
Label: Code-Review=-1