Update patch set 4
Patch Set 4:

(1 comment)

I think that the cap is the better option.  I don't think fixing the tests is enough, as the failure I include below *still* fails with 0.4.1 *and* the tests fix in the dependant patch.

My opinion is that we should take the cap as it seem like the most defensive thing to do and get to the bottom if the tests once the barbican gate is passing in kilo again.

Patch-set: 4