Bump python-senlinclient version to 0.3.0
Bump senlinclient version. Version 0.3.0 completely rewrote the client
interface and deperecated the previous one. Callers don't have to know
the resource-level abstractions any more. The only known downstream
projects are: senlin-dashboard and heat. Both projects are already
migrated to the interface exposed by 0.3.0. We have no plan to continue
support older version (<=0.2.1) which are really buggy.

On the other hand, this patch raises the upper constraint to version
0.4.0 we just released. The new version contains a full support to OSC
(openstackclient) plugins. Users are not forced to do that migration.
The calling interface remained downward compatible to that of 0.3.0.

Change-Id: Ic8c832d1c89aa5205279ab8ba8a89f673364add3