Update patch set 11
Patch Set 11:

Doug, in that case the published documentation is wrong:

https://github.com/openstack/requirements#proposing-changes

I've had people tell me they won't +1 this change without those changes done, and now you -1 with these changes. The authoritative source says to generate the upper-constraints.txt, so I want to leave it that way for now. If it's wrong, can we update the doc?

I'm just a little frustrated that I spent hours getting this working to be told it's not needed. I just need consistent information as to how to build these changes and get them approved, and right now the docs + some people are telling me different than what you are.

Patch-set: 11