add simpler instructions for adding new dependencies
Not every new dependency needs to completely recalculate the
upper-constraints.txt file. We want to know that the changes in that
file when we add something new are truly related to the new dependency,
and not just a side-effect of the update script discovering that there
has been a new release of something completely unrelated. The bot will
cover those updates separately.

Change-Id: Ieadf33401830550c6fae8f319661218071a1ad73