Update patch set 4
Patch Set 4:

Maybe you can work with the author of the duplicate patch and abandon one?

Please add the check-requirements jobs to the python-monascaclient repository before adding it to projects.txt. You'll find that you'll need to update your requirements list to get it in sync with the global requirements list, and after that is done we can add python-monascaclient to global-requirements.txt and upper-constraints.txt.

Patch-set: 4
Label: Verified=0