Update flake8 constraint for the off-by-default feature
Hacking wishes to release a new version with rules that are
off-by-default and capable of being selectively enabled. Presently the
versions of flake8 that are allowed to use this logic are not sufficient
to enable this behaviour. (2.4.x had the ability to detect plugins that
were off-by-default but no clean way to enable them given pep8's
limitations arouns --select and --ignore). 2.5.0 added a new flag to
allow users to enable off-by-default checks independently.

Change-Id: I0ee3bb69de85d9f3d5f14602016ae27b8756e53a