Make upper constraint actually be "upper" constraints
Currently, the upper constraints are using the "===" which means that
they are not upper constraint, but the actual version that pip is going
to install. It's not an _upper_ bound.

Let's use <= as an operator to be able to install lower version.

This is currently a problem in Gnocchi/Heat/Ceilometer/Aodh integration
gate as it forces keystonemiddleware===4.0.0 whereas Gnocchi wants <4
and other projects are good with >=2.3: so everybody is compatible, but
this file breaks this good intention.

Change-Id: I0fa889fd41f9a74297e7cd2ace61434e13392e5b