Add ndg-httpsclient to global requirements
This package is recommended for secure operations by:
http://urllib3.readthedocs.org/en/latest/security.html

It has a BSD license and supports both Py2 and Py3.
It is part of urllib3[secure] and requests[secure].

We would like to import this in the Octavia project so 'requests' can
properly verify HTTPS requests and validate client/server certs.

Change-Id: Id233df6061da47bd28f543615b6bf2a75e29a6cb