Update patch set 1
Patch Set 1:

Ah, I see I missed some responses...
I do not know if there is an upstream bug for this in requests. We will have to revisit that testing and see if we can get data to submit one, if it doesn't already exist -- though I thought it was well documented from urllib3 that this is the appropriate way to handle SSL verification.

I'll see about answering the rest of the questions in the submission form, but I thought packaging didn't matter for things in pip with no other system deps? (Other than pyOpenSSL which is already required globally)

Patch-set: 1