Update patch set 1
Patch Set 1:

I can't seem to repro the expired cert validation issue anymore, so it's possible that was an issue on our side that was fixed elsewhere. Looking at BREACH, I am not sure whether we fit the full vulnerability criteria, so it's possible we may not need this after all. We may want to simply ignore and squelch the warnings, re: http://urllib3.readthedocs.org/en/latest/security.html#within-the-code

Patch-set: 1