Update patch set 2
Patch Set 2:

Hey Dina,

I did see there was a change in 1.0.0 to handle staticmethod. But I can with 100% certainty state that 1.0.0 is broken with handling staticmethod in the cinder code base, and from what I've seen from other reports it is breaking other projects as well.

We need this excluded until a working release of osprofiler is available.

Patch-set: 2