Update patch set 2
Patch Set 2:

All fibre channel tests in Cinder are failing from around noon central time on the Feb 3rd from this. Here are a few runs:

https://review.openstack.org/#/c/268344/
https://review.openstack.org/#/c/276641/
https://review.openstack.org/#/c/269624/
https://review.openstack.org/#/c/260930/

The failure can be seen in the screen-c-vol logs for the third party CI test runs for "FC" or "fibrechannel" test runs (Pure, Dell, Tegile, etc.). Just search for Traceback.

The code it chokes on (in these cases) is the call in to here:
https://github.com/openstack/cinder/blob/master/cinder/volume/driver.py#L2858

That internally calls a staticmethod here:
https://github.com/openstack/cinder/blob/master/cinder/volume/driver.py#L2867

There are other similar bits of code in Cinder that I would expect to fail as well, but don't appear to be hit in these tempest runs.

Patch-set: 2