Update patch set 1
Abandoned

I think it's still possible we might be vulnerable to the CRIME exploit without this, but this is significantly mitigated by a number of other factors, like the use of private networks and secure deployment options that limit attack surface area. I'm willing to drop this for the time being, and we can revisit it if we decide this is actually a serious concern.

Patch-set: 1
Status: abandoned