Update patch set 2
Patch Set 2: Code-Review-1

In the first version of the commit message you mention copyright issues. We take licensing of our dependencies seriously, so before I approve this I would like to understand what you meant. I've read the page you linked to, but I don't understand how that relates to this library because I'm not familiar with the author or code. Can you elaborate on what you meant?

Patch-set: 2