Update patch set 2
Patch Set 2:

Python library pyeclib can use many EC backends, for example liberasure and jerasure. Because there is possibility for licensing issue with jerasure (linked page) we want to switch to liberasure by default. This change needs version bump.

I'm not lawyer, but we don't need jerasure by default, so drop it and "be on safe side".

Patch-set: 2