Update patch set 2
Patch Set 2:

Doug, we raised version due to the fact, that previous versions broke some specific gates (Cinder-drivers specific), that was not seen after usual tempest runs with the profiling feature turned off. That could be seen only after very specific gate jobs began to fail, sadly we were not able to see it from checks and tests we performed. So we fixed it and cut 1.0.1 version.

Patch-set: 2