Update patch set 1
Patch Set 1: Code-Review+1

yay bigger number!  This is what I'm packaging and it already meets the >= 1.0.8 requirement.

What is the appropriate cadence to update the requirement for an upstream package that has active contribution from openstack project maintainers?  Every time *it* releases?  Every time *openstack* releases (or milestones)?  Only in preparation for a change to swift that *requires* a newer pyeclib?  To get across 1.0.7 we had to have hacks in swift to work around the backwards incompatible change in the dependency - can we avoid that?

... there has to be some prior experience in the openstack ecosystem we can draw from, but I do like 1.2.0 and have no problem recommending that is the version of pyeclib that packagers should be building for swift.

Patch-set: 1
Reviewer: Gerrit User 1179 <1179@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1