Add "virtualbmc" dependency
This patch is adding the "virtualbmc" project to the global-requirements
file.

The project is a small utility/library that allows users to create
virtual/fake BMCs that will convert IPMI commands to libvirt commands;
this does allow the Ironic project to test the same driver used with
for the physical hardware with VMs in gate.

Currently, Ironic uses a testing driver called "ssh" (pxe_ssh, agent_ssh)
which is responsible for power controlling the VMs by SSH'ing into the
host and issuing virsh commands. This approach is not ideal because we
are not actually testing a driver that we recommend for production in
the gate.

Related-Bug: #1544642
Change-Id: I314dc131ecd5b5f3aaf3ef56d6148b6e9cb39561