Update patch set 4
Patch Set 4:

Omer, if you are proposing a new global-test-requirements.txt, then please consider helping with its implementation across all the CI infrastructure (constraints jobs, bots etc) and this review will be the final one to get in once all of that is in place. Note, that will not stop future devs of your repo to add imports to scapy in your repo which is what we are trying to avoid here.

Patch-set: 4