Update patch set 5
Patch Set 5: Code-Review-1

Is 0.3.0 backwards-compatible?

As I said in my earlier review, we try to only raise the minimum requirement when a project needs a new feature in the new release. If the library API is compatible with the older version, it's not necessary to update the minimum version just because there were refactorings.

Patch-set: 5
Label: Code-Review=-1