Change global requirements forcing cliff!=1.16.0 and 1.17.0
Both cliff 1.16.0 and 1.17.0 does not work when an empty list is provided
(Bug report on cliff: #1539770). We need to exclude both versions. Notice
that there are no 1.18.0 and 1.19.0 releases, and the problem was addressed
in 2.0.0 release, we should be fine with the exclusion.

Conflicts:
	global-requirements.txt

Related-Bug: 1539770

Change-Id: Ie77a9622d16bd02b3607fc1c9f8da20dc1ffb856
(cherry picked from commit 0ad86c4d711e4145ad3a37951949ee30bf2c7b81)